Project Overview
In this project, I used PyTorch and SageMaker to construct a complete project from end to end. The goal of this project is to have a simple web page that a user can use to enter a movie review. The web page will then send the review off to my deployed model which will predict the sentiment of the entered review.